refactor: rename solanaAuthBridge to solanaAuthStore
