+++ shared/src/auth/solanaAuthStore.ts
